Fix crash related to using block_Fifos in output module
When we got rid of the libav queues we switched to the VLC standard
mechanism, which is a block_Fifo.  However it turns up the blocks
are already on a list and aren't dequeued prior to being handed off
to the output module.  Hence, we corrupt p_next when inserting those
same blocks onto a block_Fifo owned by the output module.

Copy the block and insert the copied block onto the FIFO, which
incurs a performance penalty but avoids the p_next corruption.
The performance penalty should be relatively small (since this is
just audio), and it's no worse than when we were using libav FIFOs
which also copied the audio data.