Refactoring of 2038 decoder
There are a bunch of different fixes in this commit, but ultimately
the whole module will be squashed into a single patch so I don't
really care.

Change the lifecycle of the KL demux object to match the decoder.
Before the lifecycle was only a single MPEG packet, which would
prevent PES packets from being properly decoded if they crossed an
MPEG packet boundary.

Take into account the PTS when grouping all the lines associated
with a particular frame of video.  According to the 2038 spec, all
lines in PES packets with the same PTS should appear in the same
frame of video.

Deal with the Adtec PTS clocking being way out of sync with the
video/audio PTS clocks.  At least in the Adtec sample stream we
have been using for debug, the PTS clock is off by 47 seconds, which
effectively prevents the VANC data from being output in a timely
manner.  Compute a clock skew based on the PCR, and use that as the
time basis.  We probably should make this behavior configurable,
since it's likely it's specific to Adtec's encoder (and with other
encoders we likely would want to really use the PTS to output the
events in the correct video frame rather than immediately).

Fix a couple of obvious memory leaks.  I still need to run valgrind
against the decoder to catch any others.