Add test code to output arbitrary 104 messages
Add code which will check for the presence of a file named
/tmp/scte104-payload.bin.  If the file is found, push it out as
a properly formatted SCTE 104 message on line 11, and then
delete the file.

This patch is just for debugging, and makes it easy to do 104
output testing during playback by just copying one of a set of
precanned messages to the /tmp file.

Examples:
cp ~/scte104-payload-in-to-network.bin /tmp/scte104-payload.bin
cp ~/scte104-payload-out-of-network-30-seconds.bin /tmp/scte104-payload.bin