Add support for properly arranging VANC lines
There was a long-standing bug where the decklink output module
wasn't taking into consideration what packets might already be on
a VANC line and always outputting a VANC packet to offset zero.
This would cause VANC packets to be overwritten, whereas they should
have been appended after existing entries.

Introduce the libklvanc framework for keeping track of the VANC
packets on a given line, to ensure that all VANC entries get into
the output and that there are no gaps between VANC packets (which
would violate SMPTE ST 291-1-2011 Sec 7.3).

We also do a check for illegal byte sequences not permitted and
exclude any such VANC packets (see ST 291-1-2011 Sec 9.1 Protected Codes).

This patch not only takes care of SMPTE2038/SCTE35, but also converts
the existing local 708B and AFD functions to use the new framework,
which eliminates duplicated instances of the v210 colorspace conversion
code as well as the requirement that payloads be aligned on a 6-byte
boundary.