Add initial support for KL Colorbars input module
Hook into the libklbars library to produce colorbars
via an input plugin.

Supports specifying the width/height via module parameters,
as well as a custom line of text which will appear on line 2.

By default the module will also display a frame counter so you
can see that the video is moving, but this may be moved into
an option in the future.

Likewise, for the moment the only pattern which can be shown
is the EIA-189-A colorbars, but we will likely hook into other
patterns available in libklbars in the future (configurable
via an input module option).

Example usage:

./vlc klbars:// --klbars-width=1280 --klbars-height=720 --klbars-custom-text "LTN Global Communications"