Fix non-deterministic channel ordering for SDI output
Due to a misunderstanding of how the audio_output_t objects are
created (i.e. they are created when the first packet of data
arrives, not when the entry is enabled in the PMT), the order
in which the audio streams were associated with individual SDI
audio pairs was non-deterministic.

Have the decoder object forward it's PID when creating the audio
output, so that the audio outputs can be ordered based on their
PID.  This is not as good as ordering based on the position within
the PMT, but the PID is much easier to expose for now than the
position within the list.

This functionality can still be used in conjunction with the
channel re-mapping feature if the channels need to be reordered
in the case of 5.1 discrete audio channels being encoded over SDI.