Fix issue where 8th audio channel wasn't being created
The existing code which would make room for a new audio channel
in the middle of the list was only doing the copy from 7..n,
whereas it should have been copying from 7..n.  This resulted
in the last audio channel always being empty.

The actual bug is an artifact of the original implementation I
wrote which iterated upward from n..7, whereas we changed the
code to iterate downward from 7..n to be a bit cleaner (i.e. to
avoid having to stash the values associated with channel n+1 during
the loop).
.