Don't abort frame output if we hit an unrecognized VANC line
The existing code would bail out completely if a VANC line was
requested which isn't valid.  This would cause frozen video for
example if mis-configuring the vout-cc-line or vout-afd-line to
some invalid value.

Instead of returning out, just throw the line on the floor with an
error and continue processing.