Output VANC on correct lines when in Hp29 mode on Duo2
The scheme for specifying VANC lines on the Duo2 differs from the
original Duo card.  The original Duo card output would go on
lines 1-20, and apparently cannot output to the second field
(presumably to a driver bug).

The new Duo2 card can output to 1-20 and 564-583, but the VANC
lines provided to the drivers have to be in the range 1-41 and
the lines are interleaved between the two fields.

Hence before this change, with the Duo2, specifying CC on line
16 would actually put it on 8, and specifying line 15 would put
it on line 572.

Put some additional logic into the application such that the VANC
data actually ends up on the line specified by the user.  With
this change, the CC line, AFD line, and whatever lines provided
via the SMPTE 2038 decoder will appear on the specified line, which
can range from any line on 1-20 and 564-583.