Implement an FIFO to handle packets scheduled too far in the future
The hardware audio FIFO in decklink hardware can only hold one second
worth of audio.  Any attempt to schedule audio packets further in
the future will result in them being truncated/discarded.

Implement a FIFO that we can insert into in such cases, and push
the audio to the hardware once the packets are less than one second
in the future.

This is targeted at cases where the VLC core is delivering audio
to be played out more than a second in the future.  This includes
cases such as network streams when "--network-caching" isn't set
to some lower value (default 1000ms), as well as file playback
on OSX (due to a bug in detecting whether a filesystem is local).
It can also occur if the PCR is updated late and thus the pts_delay
is pushed to a higher value by the VLC core in an attempt to
compensate.  This includes cases where there may be packet loss and/or
a network outage on the incoming UDP stream.