Fix refcount for number of active streams
The code was apparently missing to decrement the reference count,
which would result in no audio if playing streams on a loop (i.e.
using the "-L" argument).  This was because the active_count
would never equal the ready_count.

Also, we need to start using the stop() callback to be balanced
with the start() callback, since the audio device may not actually
be closed when streams go away.  When the device was being closed
we would be ok since that would destroy all streams, but if an
individual stream is removed, the destroy wasn't being called.