Prefer interlaced formats over progressive when choosing output mode
The existing code would pick the first supported mode it encountered
in the list which matched based on resolution and framerate.  However
this ended up being 1080pSF, which causes inconsistent behavior
depending on the downstream equipment and whether the Payload Identifer
VANC data is present (it is for Duo2, but not for Duo cards).

Because the VLC framework doesn't provide us with whether the video is
interlaced, we cannot know definitively which mode to choose.  So
look for interlaced modes first and then check for progressive modes.
This should work fine for LTN given in the broadcast industry 1080p/30
and 1080p/25 are not very common.  The operator can still override
the behavior by specifying the --decklink-vout-mode= command line
argument.