Fix CC passthrough to decklink output if Bar Data present in SEI
The existing code would look for an ITU T.35 packet with a GA94
user_identifier.  However the GA94 user identifier is used for both
CC data as well as bar_data (i.e. user_data_type_code).

Because of this, if the H.264 SEI contained a bar_data packet
prior to the cc_data packet, the parser would stop when it saw
the bar_data packet and never get to the subsequent cc_data
packet.

Update the logic which looks for the CC packet to include the
user_data_type_code byte, so that we only match against GA94 packets
where user_data_type_code is 0x03 (cc_data), and in particular so
we don't also match against packets where user_data_type_code is
0x06 (bar_data).

See SCTE 128-1 2013 Sec 8.2 for the semantics of the ATSC1_data() syntax.