Fix case where decklink driver gets wedged and won't EnableVideoOutput
There is a bug either in the Duo2 or 10.8.5, where if you kill VLC
with SIGKILL the buffers don't get properly dereferenced.  The result
is any subsequent attempt to enable the video output in a higher
resolution will cause the call to EnableVideoOutput to fail.  The
bug can be worked around though by first enabling the output in
whatever the previous mode was, and then disable the output.  That
causes the driver to properly recover the buffers and then the call
to EnableVideoOutput can be made with the higher resolution

Reproducible with the following:

FILE1=~/480ifile.ts
FILE2=~/720pfile.ts

./vlc $FILE1 -I cli --vout decklinkoutput -Idummy --aout decklinkoutput
sleep 5
kill -9 `pidof vlc`
sleep 5
./vlc $FILE2 -I cli --vout decklinkoutput -Idummy --aout decklinkoutput